Split inline modules into their own .h/.cpp in Light, Led, Temperature, Ringer, BLELed
Applies the module-structure convention established in HelloIotsa (#156)
to the remaining examples that still declared their app-specific module
class inline in main<Name>.cpp: IotsaLightMod, IotsaLedControlMod (Led
and BLELed each get their own copy, same class name, different pin/BLE
behaviour), IotsaTemperatureMod, and IotsaAlarmMod. main<Name>.cpp is now
glue only: module includes, instantiation, setup()/loop().

Ringer's alarmEndTime global also moved into IotsaAlarmMod as a private
member, since it was module-private state incidentally living outside
the class.

BLEClient is deliberately left as-is -- still thin enough today that a
split isn't warranted yet (see cwi-dis/iotsa#158 discussion).

All 13 affected generated envs (nodemcuv2/esp32thing/esp32c3devkit across
the five examples) build successfully.
