+++ examples/Led/iotsaLedControlMod.h
@@ -1,0 +1,26 @@
+#ifndef _IOTSALEDCONTROLMOD_H_
+#define _IOTSALEDCONTROLMOD_H_
+
+#include "iotsa.h"
+#include "iotsaLed.h"
+
+//
+// LED module.
+//
+class IotsaLedControlMod : public IotsaLedMod {
+public:
+  using IotsaLedMod::IotsaLedMod;
+  void serverSetup() override;
+#ifdef IOTSA_WITH_WEB
+  String info() override;
+#endif
+protected:
+#ifdef IOTSA_WITH_API
+  bool getHandler(const char *path, JsonObject& reply) override;
+  bool putHandler(const char *path, const JsonVariant& request, JsonObject& reply) override;
+#endif
+private:
+  void handler();
+};
+
+#endif
